KSENROLL-14743 Setup initial pull-request-builder pom
The per pull request CI uses this project which holds the git-workflow
maven plugin configurations for driving the process.

The pull-request-builder will checkout the ks-development and
ks-development-impex projects and build them as required.

As each phase is completed the Github Commit Status Api will be used to track
state and also allow jobs to be resumed if there are not pending or failed jobs.