KSENROLL-14743 Add scripts to identify sql change and impex commits
Instead of using the Github API or using JGit I have created two shell scripts
that will let us identify which pull request commits have sql changes and
if the latest of those is represented in the impex side.

identify-ks-repo-sql-changes.sh operates on the target/ks-repo to walk
from the pull request head commit back to the base.

identify-if-ks-impex-is-up-to-date.sh uses the commit id returned by
identify-ks-repo-sql-changes.sh to see if it matches the commit id
in the latest pull request branch commit on the impex side.

If it matches no impex is required.

If it is not found manual impex is required.

Work on modularizing the scripts for the main pull request processing scripts.